predator12-local/backend/requirements.txt
# Core FastAPI and Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.0
aiofiles==23.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# HTTP Clients
httpx==0.25.2
requests==2.32.3
aiohttp==3.9.1

# Observability & Monitoring
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-exporter-otlp==1.25.0
structlog==23.2.0

# Database & Storage
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==4.6.0
opensearch-py==2.6.0
qdrant-client==1.7.0

# Task Queue & Workers
celery[redis]==5.3.4
rq==1.16.2

# Message Brokers
aiokafka==0.10.0
asyncio-mqtt==0.16.1

# Data Processing & ML
pandas==2.2.2
numpy==1.24.4
scikit-learn==1.3.2
mlflow==2.8.1

# Configuration & Utilities
PyYAML==6.0.2
pydantic==2.5.2
networkx==3.2.1
gitpython==3.1.40

# AI Model Providers
openai==1.12.0
anthropic==0.8.1
cohere==4.37
mistralai==0.1.8

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
